+++ DSCResources/WindowsClient/WindowsClient.schema.psm1
@@ -1,0 +1,145 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+using module ..\helper.psm1
+using module ..\..\PowerStig.psm1
+
+<#
+    .SYNOPSIS
+        A composite DSC resource to manage the Windows 10 STIG settings
+
+    .PARAMETER OsVersion
+        The version of the server operating system STIG to apply and monitor
+
+    .PARAMETER StigVersion
+        Uses the OsVersion and OsRole to select the version of the STIG to apply and monitor. If
+        this parameter is not provided, the most recent version of the STIG is automatically selected.
+
+    .PARAMETER ForestName
+        A string that sets the forest name for items such as security group. The input should be the FQDN of the forest.
+        If this is omitted the forest name of the computer that generates the configuration will be used.
+
+    .PARAMETER DomainName
+        A string that sets the domain name for items such as security group. The input should be the FQDN of the domain.
+        If this is omitted the domain name of the computer that generates the configuration will be used.
+
+    .PARAMETER Exception
+        A hashtable of StigId=Value key pairs that are injected into the STIG data and applied to
+        the target node. The title of STIG settings are tagged with the text ‘Exception’ to identify
+        the exceptions to policy across the data center when you centralize DSC log collection.
+
+    .PARAMETER OrgSettings
+        The path to the xml file that contains the local organizations preferred settings for STIG
+        items that have allowable ranges.
+
+    .PARAMETER SkipRule
+        The SkipRule Node is injected into the STIG data and applied to the taget node. The title
+        of STIG settings are tagged with the text 'Skip' to identify the skips to policy across the
+        data center when you centralize DSC log collection.
+
+    .PARAMETER SkipRuleType
+        All STIG rule IDs of the specified type are collected in an array and passed to the Skip-Rule
+        function. Each rule follows the same process as the SkipRule parameter.
+#>
+Configuration WindowsClient
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateSet('10')]
+        [string]
+        $OsVersion,
+
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [ValidateSet('1.14')]
+        [version]
+        $StigVersion,
+
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [string]
+        $ForestName,
+
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [string]
+        $DomainName,
+
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [psobject]
+        $Exception,
+
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [psobject]
+        $OrgSettings,
+
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [psobject]
+        $SkipRule,
+
+        [Parameter()]
+        [ValidateNotNullOrEmpty()]
+        [psobject]
+        $SkipRuleType
+    )
+
+    ##### BEGIN DO NOT MODIFY #####
+    <#
+        The exception, skipped rule, and organizational settings functionality
+        is universal across all composites, so the code to process it is in a
+        central file that is dot sourced into each composite.
+    #>
+    $dscResourcesPath = Split-Path -Path $PSScriptRoot -Parent
+    $userSettingsPath = Join-Path -Path $dscResourcesPath -ChildPath 'stigdata.usersettings.ps1'
+    . $userSettingsPath
+    ##### END DO NOT MODIFY #####
+
+    $technology        = [Technology]::Windows
+    $technologyVersion = [TechnologyVersion]::New( $OsVersion, $technology )
+    $technologyRole    = [TechnologyRole]::New( 'Client', $technologyVersion )
+    $stigDataObject    = [StigData]::New( $StigVersion, $OrgSettings, $technology,
+                                          $technologyRole, $technologyVersion, $Exception,
+                                          $SkipRuleType, $SkipRule )
+    
+    #### BEGIN DO NOT MODIFY ####
+    # $StigData is used in the resources that are dot sourced below
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseDeclaredVarsMoreThanAssignments",'')]
+    $StigData = $StigDataObject.StigXml
+
+    # $resourcePath is exported from the helper module in the header
+
+    # This is required to process Skipped rules
+    Import-DscResource -ModuleName PSDesiredStateConfiguration -ModuleVersion 1.1
+    . "$resourcePath\windows.Script.skip.ps1"
+    ##### END DO NOT MODIFY #####
+
+    Import-DscResource -ModuleName AuditPolicyDsc -ModuleVersion 1.2.0.0
+    . "$resourcePath\windows.AuditPolicySubcategory.ps1"
+
+    Import-DscResource -ModuleName AccessControlDsc -ModuleVersion 1.1.0.0
+    . "$resourcePath\windows.AccessControl.ps1"
+
+    Import-DscResource -ModuleName PolicyFileEditor -ModuleVersion 3.0.1
+    . "$resourcePath\windows.cAdministrativeTemplateSetting.ps1"
+
+    Import-DscResource -ModuleName WindowsDefenderDSC -ModuleVersion 1.0.0.0
+    . "$resourcePath\windows.ProcessMitigation.ps1"
+
+    Import-DscResource -ModuleName PSDesiredStateConfiguration -ModuleVersion 1.1
+    . "$resourcePath\windows.Script.wmi.ps1"
+    . "$resourcePath\windows.WindowsFeature.ps1"
+
+    Import-DscResource -ModuleName xPSDesiredStateConfiguration -ModuleVersion 8.3.0.0
+    . "$resourcePath\windows.xService.ps1"
+    . "$resourcePath\windows.xRegistry.ps1"
+
+    Import-DscResource -ModuleName SecurityPolicyDsc -ModuleVersion 2.4.0.0
+    . "$resourcePath\windows.AccountPolicy.ps1"
+    . "$resourcePath\windows.UserRightsAssignment.ps1"
+    . "$resourcePath\windows.SecurityOption.ps1"
+}