+++ Tests/Unit/DSCResources/Composite.tests.ps1
@@ -65,7 +65,6 @@ Describe 'Common Tests - Configuration Module Requirements' {
         }
     }
 
-
     Describe 'Composite Resources' {
 
         $manifestDscResourceList = $Manifest.DscResourcesToExport
@@ -80,12 +79,18 @@ Describe 'Common Tests - Configuration Module Requirements' {
         $TechnologyRoleFilter = @{
             Browser          = 'IE'
             DotNetFramework  = 'DotNet'
+            IisServer        = 'IISServer'
+            IisSite          = 'IISSite'
+            OracleJRE        = 'OracleJRE'
             SqlServer        = 'Database|Instance'
             WindowsDnsServer = 'DNS'
             WindowsFirewall  = 'FW'
             WindowsServer    = 'DC|MS'
             Office           = 'Outlook2013|Excel2013|PowerPoint2013|Word2013'
+            WindowsClient    = 'Client'
+            FireFox          = 'FireFox'
         }
+
         Foreach ($resource in $moduleDscResourceList)
         {
             Context "$resource Composite Resource" {
@@ -119,4 +124,3 @@ Describe 'Common Tests - Configuration Module Requirements' {
     }
 }
 #endregion Tests
-