+++ Tests/Unit/Module/Convert.IISLoggingRule.tests.ps1
@@ -170,17 +170,17 @@ try
                 $classMethods = @('SetLogCustomFields', 'SetLogFlags', 'SetLogFormat', 'SetLogPeriod'
                     'SetLogTargetW3c', 'SetStatus')
 
-                foreach ( $method in $classMethods )
+                foreach ($method in $classMethods)
                 {
                     It "Should have a method named '$method'" {
-                        ( $rule | Get-Member -Name $method ).Name | Should Be $method
+                        ($rule | Get-Member -Name $method).Name | Should Be $method
                     }
                 }
 
                 # If new methods are added this will catch them so test coverage can be added
                 It "Should not have more methods than are tested" {
                     $memberPlanned = Get-StigBaseMethods -ChildClassMethodNames $classMethods
-                    $memberActual = ( $rule | Get-Member -MemberType Method ).Name
+                    $memberActual = ($rule | Get-Member -MemberType Method).Name
                     $compare = Compare-Object -ReferenceObject $memberActual -DifferenceObject $memberPlanned
                     $compare.Count | Should Be 0
                 }
@@ -188,7 +188,7 @@ try
         }
         #endregion
         #region Method Tests
-        foreach ( $rule in $rulesToTest )
+        foreach ($rule in $rulesToTest)
         {
             $checkContent = Split-TestStrings -CheckContent $rule.CheckContent
             Describe 'Get-LogFlag' {