+++ CHANGELOG.md
@@ -2,17 +2,36 @@
 
 ## Unreleased
 
+## 2.2.0.0
+  
 * Added the following STIGs
+  * IIS Site 8.5 STIG V1R2
+  * IIS Site 8.5 STIG V1R3
+  * Oracle JRE 8 STIG V1R5
   * Microsoft Outlook 2013 STIG V1R12
   * Microsoft PowerPoint 2013 Stig V1R6
   * Microsoft Excel 2013 STIG V1R7
   * Microsoft Word 2013 STIG V1R6
-* Added the DSC Composite Resource for Microsoft Office 2013 STIGs
-* Added required module "PolicyFileEditor" for applying Office HKey_Current_User settings
 
-* Updated the following Stigs
-  * WIndows 10 STIG - Updated for xRegistryResource and RegistryRule to Manual Rule
-  
+* Added the following DSC Composite Resources
+  * Microsoft Office 2013 STIGs
+  * FireFox STIG
+  * IIS Site STIG
+  * IIS Server STIG
+  * Oracle JRE STIG
+  * Windows10 STIG
+
+* Newly required modules
+  * PolicyFileEditor
+  * FileContentDsc
+  * WindowsDefenderDSC
+  * xWebAdministration
+  * xWinEventLog
+
+* Updated required module versions
+  * xDnsServer from 1.9.0.0 to 1.11.0.0
+  * SecurityPolicyDsc from 2.2.0.0 to 2.4.0.0
+
 ## 2.1.0.0
 
 * Migrated Composite resources to the xRegistry resource